Fixing mockito exception
